requirements.txt
discord.py>=2.3.2
python-dateutil
pytz
dateparser
aiosqlite